# test_linspace : remove explicit for-loop